+++ front/src/components/organisms/CardResource.tsx
@@ -85,6 +85,7 @@ const CardResource = ({
             voteCount={voteCount}
             resourceId={id}
             handleAccessModal={handleAccessModal || undefined}
+            fromProfile={fromProfile}
           />
         </CounterContainerStyled>
       )}